feat [SRTP-725]: refactoring
